refactor(forge): split Forgejo issue methods into forgejo_issues.go (#1025)
